Improved country example

+++ examples/05.recycler-view/CountryExplorer/app/src/main/java/qa/edu/cmps312/countryexplorer/model/Country.kt
@@ -24,6 +24,6 @@ data class Country (
     val capital: String,
     val continent: String,
     val region: String,
-    val population: Int,
+    val population: Long,
     val area: Long = 0
 )
